move the tests that require personium-engine settings to a special package
